Rewrite using GetLinkColours hook.
The old version slurped a bunch of data and wasted a lot of cache
to store that.  This new version just asks the DB for categories
once per request by using the hook that lets us set all the internal
link colors at once.  Should be faster, even without caching.